chore: move custom manager to a new file
